+++ t/01-basic.t
@@ -8,14 +8,14 @@ use Time::Piece;
 my $tl = SVG::Timeline->new;
 
 $tl->add_event({
-  start => 1987,
-  end   => localtime->year,
+  start => '1987',
+  end   => '2022-10-23',
   text  => 'Perl',
 });
 
 $tl->add_event({
-  start => 2017.5726,
-  end   => localtime->year + ( localtime->yday + 1 ) / ( localtime->is_leap_year ? 365: 366 ),
+  start => '2017-07-29',
+  end   => '2022',
   text  => 'SVG::Timeline on CPAN',
 });
 
@@ -29,7 +29,7 @@ $tl->add_event({
   end   => localtime->year,
   text  => 'Python',
 });
-is($tl->count_events, 2, 'Correct number of events');
+is($tl->count_events, 3, 'Correct number of events');
 
 my $vb2 = $tl->svg->getChildren->[0]{viewBox};
 isnt($vb1, $vb2, 'Viewbox changed');